+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 * added `file_contents` schema 
 * added nested configuration sources
 * added `extract_optional`
+* added some syntactic sugar for the source dict
+* now unparseable "type" entries will be returned as-is
+* added `ConfigurationMisconfiguredError`